引入mapper通用模板类

+++ sample-service/src/main/java/com/sample/service/LogInfoService.java
@@ -4,7 +4,7 @@ import com.sample.model.LogInfo;
 
 public interface LogInfoService {
 
-    LogInfo selectById(Integer id);
+    LogInfo selectById(Long id);
 
     void insert(LogInfo info);
 }
