+++ sample-service/src/main/java/com/sample/service/UserInfoService.java
@@ -3,7 +3,7 @@ package com.sample.service;
 import com.sample.model.UserInfo;
 
 public interface UserInfoService {
-    UserInfo selectById(Integer userId);
+    UserInfo selectById(Long userId);
 
     void insert(UserInfo info);
